CDE-67 feat: Refactor out chip component
